seperate js into different files
