+++ src/pages/singleEvent/helperFunctions.js
@@ -1,0 +1,21 @@
+/* HELPER FUNCTIONS  */
+/* Get ID from window */
+export const params = new URLSearchParams(window.location.search);
+export const meetupId = params.get("id");
+
+export function formatDate(data) {
+  const date = new Date(data);
+  return date.toLocaleDateString("no-NO", {
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+  });
+}
+
+export function formatTime(data) {
+  const time = new Date(data);
+  return time.toLocaleTimeString("no-NO", {
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+}
