--- src/pages/singleEvent/singleEvent.ts
@@ -1,4 +1,0 @@
-const postBtn = document.getElementById("post-btn");
-const postOverlay = document.getElementById("post-overlay");
-
-postBtn.addEventListener("click", () => {});
